client API: Synchronous Element API proposal
Eric Hu came up with a good idea for composing Elements and
ElementLinks. Clients of route-rs will implement an Element trait, which
requires just a "process" function where they specify what inputs and
ouputs the Element transforms. The route-rs library has a ElementLink
that calls an Element.process and links elements to each other, probably
through a macro or the graph visualizer.