Immediate Notify: Use task::current().notify()
In our AsyncElement, we previously used a ForeverDrain element that
wouldn't respond to Async::NotReady, and therefore would continue to
poll the functions non-stop. Now we switch out this functionality by,
when the poll() is about to return an Async::NotReady, rescheduling the
task with Tokio by calling task::current().notify() inside the poll
function.

An aside about Tasks, a task is a unit of abstration over a bunch of
related futures. Tokio has this concept of scheduling tasks, done
through the notify function, which tells the Runtime to try calling
the topmost poll function of whatever Task has had it's notify
function called recently. Think of it like reschedule.

In this case, we call notify before we even return the NotReady,
which is an improvement from the previous use of the forever
drain. Since we yield our computation momentarily, it is still less
than ideal, as we are not driving the notify function using some
notion of when there is new work to complete. This will be
fixed in later efforts.

Some additional reading I found helpful on this matter:
https://docs.rs/futures/0.1.28/futures/task/struct.Task.html
https://cafbit.com/post/tokio_internals/