Impl Future for AsyncElementLink.frontend
Previously we implemented Stream for both the frontend and backend
portions of the AsyncElementLink. This makes some intuitive sense,
since a Stream is meant to stand for a continuous stream of futures,
which represents the sequence of packets being processed by an
element fairly well. However, Tokio's executor needs handles to Futures
rather than to Streams; we had to place a shim, such as the
ExhaustiveDrain utility element, in front of each stream before they are
provided to the runtime. In order to simplify the syntax, and reduce the
complexity of our API, I implemented Future for the frontend, so that
the frontend can be handed to the runtime directly.

In this implementation of Future, it never returns Async::Ready until
the element is ready to be torn down. An Element's frontend will continue
to process packets off its input stream in a loop until it can no longer
do work, in which case it will return Async::NotReady.

Note that processing logic has not actually changed, since this was the
functionality caused by placing the frontend stream into the
ExhaustiveDrain utility element.

Relevant tests were also adjusted, all tests pass after adjustment.