Add packet_generators::immediate_stream as a wrapper around iter_ok
'immediate_stream' is essentially a wrapper around
'futures::stream::iter_ok', which transforms an IntoIterator into a
Stream which yields an item on each call to 'poll', and never returns
Async::NotReady. It's useful to abstract this kind of Stream into
util::test since it's used in multiple tests already, and allows the
tester to focus on testing behavior by supplying known facilities.