packet_collectors.rs:  Impl Exhaustive Colllector
Exhaustive Collector is a packet collector that will grab every packet
out of a stream, but it is different from Exhaustive Drain in that it
writes the collected packet into the provided channel handle. This is
useful because the test may then inspect the channel to determine
whether the test was a pass for fail.