packet_generators.rs: Impl PacketIntervalGenerator
This commit adds the PacketIntervalGenerator, a generator that
takes an interval and an object that implements Iterator, and pulls
packets from that object until it is exhausted, provided the packets
out of a Stream. The packets are produced on a fixed interval so that
the Stream sometimes returns NotReady,  to better test the router.