lib.rs: Add Collected Yield Tests
Up unti this point, our testing suite did not have a way to assert that
the output of the router matched some sort of specification.  The
tests pass or fail simply based on whether all of the futures return.

This commit puts the use the PacketIntervalGenerator and the
ExhaustiveCollector to create a suite of tests that the assert that
the series of packets that were fed into the router match the
output of the router.

Run the suite with the command `cargo test collected_yield`