Test: Reduce verbosity of successful tests
It isn't considered good form for tests to print when they are not
throwing some kind of error.  With all the prints in our testbed it can
be hard to read what passed and what failed. Made this better by
commenting out all the debugging print statements. I figure that
when there is a failed test it will be convienient to have those there
to be uncommented.  After this commit tests are much cleaner and
running `cargo test` is functional.