Implement ClassifyElement
This commit implements the ClassifyElement, a generic Asynchronous
element that takes in a packet, and sends the packet out of a port
based on a user provided classify function. In this basic implementation,
there a few less than ideal cases.

The classify function, right now, outputs a tuple with a number
determining what port to send the packet out of, perhaps this could be
a different mechanism if we so desire.

The Consumer, which takes in a packet and forwards it, via a channel,
to the appropriate provider, will stop processing packets if any one of
the channels to a port becomes full. This may be the implementation
that we want moving foward, but we should discuss this. Additionally,
at the beginning of each poll we check all the channels to see if any
are full, and await the producer of the full task.  Iterating over every
channel on each poll is not very efficient, perhaps we could do a full
buffer check at the tail, when we push a packet, and sleep then
instead, and only that the appropriate channel is less full on wake,
saving us some cycles. However, if the number of ports is low, this
cost may not be a big deal, and we can leave the function as is.