Implement ClassifyElement
This commit implements the ClassifyElement, a generic Asynchronous
element that takes in a packet, and sends the packet out of a port
based on a user provided classify function. In this basic implementation,
there a few less than ideal cases.

The classify function, right now, outputs a tuple with a number
determining what port to send the packet out of, perhaps this could be
a different mechanism.

The Consumer takes in a packet and forwards it via a channel to the appropriate
provider. It will stop processing packets if any one of the channels to a port
becomes full, because we will not have a place to store the packet after processing
if it is destined for a full channel.