AsyncElement: Test cleanup
This commit removes a bunch of extranous tests from the AsyncElement
test suite. The remaining tests now all serve a unique purpose, and are
renamed to describe what they are testing, rather than how they are
testing. In general, this change also means that all tests are now
"collected", and assert that the correct packets reach their
destination.