ClassifyElement: Add Overseer
This commit adds the overseer. It solves the same problem as the
previous commit which adds the overseer to the AsyncElement. In
this case it works slightly differently, since we need to keep
track of the number of providers that are still alive, so that
the overseer does not return until all of the provider channels
are empty, and the consumer has dropped. Then it is safe for the
overseer to cease operation. This is a bit roundabout, but without
the explicit check in crossbeam_channels for disconnection, this is
a fine proxy.