task_park: extract common utilities
It turns out that there are very repeatable patterns to dealing with
task-wake management in concurrent elements. Those are:

1) unpark_and_notify: "I have work to, so don't bother waking me up."
2) park_and_notify: "I don't have any work to do. Let me know when I do."
3) die_and_notify: "I'm shutting down, and I'll either let you wake me
up if you're around, or I'll self-notify if I die".

park and unpark variants will notify a task if it exists in the park as
the new state is swapped in, and will maintain the "dead" state so
die_and_notify works appropriately.