runtime: Move identity_elements to element module
The identity elements are useful outside just the confines of tests in
runtime. We should export them (and in the future probably things like
TraceElement and BlackHoleElement), through the element module.

Additionally, remove the id field on the IdentityElements because it is
never used and was triggering dead code warnings. Derive Default on the
IdentityElements now that they have no parameters to new(), even though
in autogenerated calls we will still always invoke the constructor for
consistency with elements that do have constructors. Remove an extra
reimplementation of IdentityElement that was in the classify_link test.

This commit specifically will be useful as I build out the examples/
crates so I don't have to reinvent the wheel so much.