Implement pipeline and provide an example
runtime::pipeline defines how a full pipeline can be built to run in a
binary crate. This means we can actually have a full router crate, so
add an example of that.

Right now this is all manually coded, but in the future the pipeline
implementations will be mostly generated from graph files.