task_park: Prevent overscheduling join_link task
In previous implementations of Task Park, we had an issue where we could
not hand out a common task handle to multiple task_parks when one task
wished to await several tasks that provide work. This was particularly a
problem in join_link, when the downstream provider needed to await any
of the consumers enqueuing a packet. This solution adds the
IndirectParked state, where the task can hand out a thread-safe
reference to a task_handle. This solves the overscheduling problem since
only the first consumer to access the reference will receive the task
handle. Subsequent accessors will receive a None, which informs them
that they do not need to awaken the provider.