Change names: Ingressor, Egressor, [type]Link
This commit changes the names of consumer to Ingressor, of provider
to Egressor, and removes the word Element from between the type of
link and Link.  So now we have ClassifyLink instead of
ClassifyElementLink.

Overall this makes the names much more clear, per Ben Weis's idea, where
the general packet flow is less confusing.