Rewrite ClassifyLink to work better in pipelines
This commit rewrites ClassifyLink to work more similarly to the
standard SyncLink. This allows us to define the dispatch from it inline
in the pipeline, which makes code generation easier in graphgen.

This commit also restructures ClassifyElement as a subtrait of Element,
which keeps the structure more consistent for our users.

examples/dns-interceptor shows the usage of ClassifyLink and JoinLink,
so we know that they mesh with the overall pipeline implementation. A
future commit will support this example with graphgen.

Fixes #77