Add checks for dependency license compatibility
Since route-rs is distributed under the MIT license, we want to make
sure that all of our dependencies are similarly nonrestrictive.
Our users will have to treat our package like the most restrictive
license in our dependency tree when deciding if it is usable in their
projects, so we should make this an easy decision for them.

This commit uses cargo-deny to enforce licensing, and updates our own
crates to properly record license information. We also have to
whitelist LICENSE-THIRD-PARTY in crossbeam-channel because it confuses
the scanner. I have manually inspected that license and it is just a
concatenation of licenses that are in our whitelist.