afpacket: low-level crate for AF_PACKET
This patch adds a new crate implementing low-level safe bindings to
AF_PACKET. The goal of this crate is to create a misuse-resistant
binding to the underlying AF_PACKET library that can be used from safe
Rust.

This crate defines a `Socket` type that is loosely based on the stdlib's
`UdpSocket` type.