Link: Add BlackHole
This is a simple link that  takes in an input stream and drops any
packets that it recieves. Much like the async_link, it will continue
to pull packets until it receives a NotReady, in which case it will
propogate the NotReady until there are further packets to blackhole.

In general this link will always be handed directly to the runtime.

This commit also moves ClassifyEvenness and ClassifyFizzBuzz to a
common element file called classify_test_elements.rs. This is done
to prevent redeclaration of ClassifyEveness in blackhole_link.rs