WIP: Initial ICMP utility to generate error messages
This is a first pass at an ICMP error message generator, and is IPv4
only for now. smoltcp does not offer a way to create all the ICMP error
types we need, so the interface is a bit complex. Some parts are
slightly hacky as well due to other smoltcp limitations.