ClassifyLink: Break Classifier away from Element
The Classifier trait is significantly different from Element such that
it should be an independent trait from it, rather than shoehorning
these two operations together. This allows us to nicely look at packets
and determine their class with an immutable reference, so we guarantee
the packages exit ClassifyLink untransformed.

Hat tip to ddowl for pushing me to make this improvement.