WIP: first attempt at packet wrapper
This is a first attempt at a packet wrapper that owns its own Vec<u8>,
so that when a packet comes in as a layer 2 frame, we can reuse this
already allocated memory, rather than re-allocating memory when
re-adding the layer 3 and layer 2 headers.

This is only a draft, seeking feedback. I believe the l2_encap methods
are of particular note in Ipv4Wrapper and Ipv6Wrapper, as well as the
layer3_(un)checked and layer4_(un)checked methods. There is also a ton
of copy and pasted code, I'm sure there's some way to re-use a lot of it
and cut down on the identical blocks.