Packet: A route-rs packet library
This is the base commit of the route-rs packet library. It is at a
point where it is not ready to be merged to master, but is ready to
being accepting PRs and outside help. Right now, there exists a unique
structure for each relevent packet type, checked constructors for all
of them, and conversion functions everywhere it makes sense. The code
is not well tested, and does not yet conform to large sections of the
RFC requirements. This is why it is living in seperate branch for now.

Until it is ready to submit to master for an intensive CR, I am going
to install myself as the BDFN (Benevelent Dictator for now) and accept
PRs solving the various issues detailed in the sprint. I will CR all
the changes myself. I am trying to move fast on this library since there
is a lot of ground to cover to get it to a functional stage, as such I
will only sometimes request CR/outside review for various changes I
make. Others will have a change to voice concerns and advocate for
changes either through submitting PRs, or during the CR process for
submitting to master.