SyncLink: impl Link
In addition, prefers the name "in_streams" since it's synonymous with
"input_streams" and "ingress_streams". Removes graphgen examples from
top-level crate so they aren't compiled while API is in flux. Modifies
the `ingressor` function to consume `self`, as builder methods always
yield a new version of `self`. This is useful when moving the Link to
a new state (see `SyncLink::ingressors` and `SyncLink::element`).