Packets: Adjust structs to carry offset data
Adds values to each struct so that we retain offsets for the start of
each layer's header. This is useful when creating a less specific
headeer from a more specific one. For instance, when we move UDP->Ipv4.
It is easy in the from function to reference segments packet_offset
field. This means that creating a segment from scratch requires
providing other values, like the packet offset. But this helps the code
remain agnostic to the various layer's thicknesses.