Packets: Add helper function to determine segment
Add two helper functions for Ipv4 and Ipv6 to determine the IpProtocol
type of their payload. These functions are useful since they can be
called when creating a new segment, TCP or UDP, so that the new function
may check that the provided data does in fact say it contains the
appropriate transport type.