AsyncLink: Move to builder pattern
Moves AsyncLink to builder pattern, and adds some additional required
trait bounds to element. This has to do with things requiring to be Send
when being put into a channel, and with the traits the compiler is no
longer able to infer that the E::Input and E::Output are Send. So we
state that as a fact.