BlackholeLink: Switch to Element from Packet
Previously, we had an issue that the we were making Blackhole link generic
off the Packet type that is flowing into it as input. This instictually makes
sense, but due to our need to create TokioRunnables with 'static lifetimes, it
means that the Packet would have to be 'static. A way to get around this is to
use an IdentityElement which is 'static