Links: Rework asserts to be more clear
Rework asserts to be more consise and clear, and remove a few that
were not needed. Also allow user to create CloneLinks of branch amount 1,
this is silly for them to do but there is no reason to stop them.  Also
use the variable num_egressors instead of branches to be more clear for the
reader.