Packets: Ethernet: Fix EtherType bug
Previously, we were not handling the case where the field proceeding the
two MAC addresses, generally called the EtherType field, is valid as
either the payload_len or an EtherType value. Which it is depends on the
value present. Generally, in modern times, it is used to indicate that
the frame contains an IPv4 packet.

We now handle this case as simply as possible, with a check that the
field is inside a valid range, and checking for valid length fields
where it makes sense.

We also change the set_payload function where it only updates the
ether_type field if it currently is being used as a payload_len. In the
future, I suspect this area will need to be built out more to support
VLAN tagging, but that is pretty far down the road.