graphgen: Generate imports with syn
The syn crate provides a bunch of nice interfaces for working with the
Rust AST. This is much more reliable and scalable than hand-rolling our
own code generator.

The main downside with working with syn is that it assumes that we have
parsed our AST from somewhere. Since we're just making it up, we don't
have legitimate Span objects corresponding to the source file locations,
so we just say everything is from the call_site. So we don't have to
write this out all over the place, bury those span creations in the
codegen module.

As a bonus, I don't have to spend nearly as much time fighting with
String and &str type conversions, hooray!