Packets: Change/unify member variable name schema
Modify the member variables of each packet type to use the same naming
schema, and cleanup variables that are not going to be used, such as the
valid_checksum marker. In the future, we want to support an actual
validated packet type, where it is in effect "ready" to be sent out of
an interface. We don't want to keep track of too much internal state in
the packet struct itself though, since it implies to the user that it
can be relied on. This is very much a flat library, intended to be
flexible to the user, rather than a strict library on rails. We are
going to allow invalid state, etc etc. It will be up to the user of the
library to validate the packets manually for the time being.

The change to keeping track of layer2 throughout the stack is also part
of an intiative to allow the use of "headless" data pieces, where higher
layer data structs can be built first without the appropriate header.
This should aid the construction of packets inside the router, where the
header for the Ethernet is not often known when the layer3 destination
is deteremined. We don't want to force the user the construct a bogus
layer2 header. Instead we will return an error on trying to convert a
higher layer packet into a lower layer one, if the layer does not exist.