Packets: make lower layer headers optional
This commit makes lower layer headers optional, so that in the future
users may create packets of a higher layer that do not have lower level
headers. This makes sense sometimes in the router, when a processor only
can construct the higher layer, and then will not have to create a dummy
lower layer header. That is doable by using the New functions.