Runtime: Fix multithread dropping bug
When running the router in a multithreaded mode, we had some race
conditions in the Drop functionality leading to weird errors. This is
resolved by moving the drop signaling out of the Drop functon, and into
the poll call. Generally these errors manifested as a Disconnected error
in the to_egressor channel. Likely this is due to some condition where
both sides of the channel are dropping at once, leading to racey
behavior. Since the poll has good assertion that the channels are
in a valid state before it pulls a packet, it can use that assertion to
signal the channel, and therefore the egressor, that it is time to drop
without running into a weird disconneted error.