codegen: Add fake_span helper
This helper is used wherever we have to generate a proc_macro2::Span
to satisfy the syn types. Since we aren't actually parsing from
anywhere, this is all made up, and using the helper is more explicit
about that so that future developers don't think these Spans have a
useful meaning.