graphgen: use syn to generate fn run
This was kind of tricky, because it's important for us to have newline
separations between the different parts of the function body in order
for it to be readable by mortals. Unfortunately, syn doesn't believe in
vertical whitespace at all.

The solution is to create a magic macro called graphgen_magic_newline!,
which we inject wherever we need a newline to appear. After syn renders
out the code it thinks is right, we can replace the macro with "\n\n",
putting a blank line in the middle of the source. After rustfmt, it
looks the way we wanted.

The other addition here is a method to generate Paths, potentially with
angle bracketed arguments on them, to make type signatures easier. This
is a bit more flexible than simple_path, and I'll reconcile them in a
future commit.