MVR: Join and dispatch by Interfaces
This takes another step towards the middle of the router, unifying the
input streams into one so we can add the packet processing pipeline.

One interesting thing about this is that we need to implement
BlackHoleLink after all, because something needs to poll for handling
packets that have somehow gotten through without any tagging. It would
be nice to be able to prove that we can't get None at the end of this
using the type system, but I'm not sure how to do it without requiring
that we make all egress port routing decisions in the same place, which
seems like a really annoying limitation. I have therefore opted to
black hole any packets which make it all the way through in this state,
because crashing the router on them feels too harsh.