MVR: Stub out protocol classifier
This will allow us to split into different processing chains for the
protocols that we are supporting in this router. All other types of
packets are dropped because we don't know what to do with them.

Note also that the classification logic is unimplemented for now. We
just treat everything as unknown, so all data will be dropped.