Runtime: ClassifyLink: Allow dropping packets
In the past, ClassifyLink was unable to drop packets. This means that if
classify is being used as a complex filter (say sorting valid IP
traffic), it was unable to drop packets. It would have to send packets
to be dropped out a specific port, which would lead to a blackhole link.
Since the BlackholeLink was removed in favor of allowing processors to
drop packets, we should also allow Dispatchers (The Fn that maps a
classificiation to a port number) to drop packets. We do this by
changing the return of a Dispatcher to be an Option type. If it wishes
to drop a packet, it returns a None, rather than Some(portnum).
ClassifyLink then drops the packet in question.