Runtime: runner() now takes a Trait Object
Before we were taking an function object that could produce us a Link
that could be run in the tokio runtime, instead now we take a Trait
Object that implements LinkBuilder. Since it implements LinkBuilder, we
can assume that the object passed in is already initialized, and we then
build the object inside the Tokio runtime context by calling
build_link(). This is much more natural than the previous
implementation, and helps to eliminate some boilerplate.