runtime: Runner framework for external links
build_and_run_router allows us to pass in external connectors to
assemble a router. This is useful so that we can be generic over the
way that packets come in and out, allowing us to use the same router
struct for production (where we hook up to real interfaces) and tests
(where we hook up to pcaps).

IngressLinkBuilder and EgressLinkBuilder provide traits that we can
conform to for various i/o links in the future.