Runtime: Update run_link to new test_link
This new testing harness allows the user to optionally specify a
timeout. If no timeout is given, test_link will run the link until it
exits. If a timeout is specified, and the link is still running when the
time expires, the link is torn down and the packets that were egressed
from the link are collected and passed to the user. If the link finishes
before the timeout, then it functions as if no timeout were specified.

This is useful for links that do not have an exit condition. The tester
can now let the links run for a set amount of time, and then check that
all the packets were transferred in time.