PHR: Init example, create main and router
This commit initializes our new Pi Home Router example in the repo,
creating a new cargo crate. It lays out a basic filesystem in a manner
that I think will be useful to the implementors working on the router
internals. The main.rs and router.rs files introduce a limited skeleton
to construct a router. It shows some basic argument parsing, and running
the router with the runner() harness provided by the runtime. The router
object in router.rs does nothing as of yet, and returns an empty Link
object when built.

There are no integration tests created yet for the router, those will
get handled down the line when packet interface intake and exhaust are
implemented.