PHR: Introduce InterfaceCollect and InterfaceDispatch
InterfaceCollect is a new link used to collect packets from all the
input interfaces and join them into a single stream. As it does this, it
also applies an annotation to each packet denoting which interface said
packet originated from. As such, it maps `EthernetFrame` to
`InterfaceAnnotated<EthernetFrame>`.

InterfaceDispatch is a new link that takes in a single stream of
interface annotated packets, and dispatches each packet to the
appropriate outbound interface, removing the interface annotation as it
does so. The outbound_interface annotation is a field of the
InterfaceAnnotated<Packet> structure, and any packet that does not have
a marked outbound interface will be dropped. As such, it maps
`InterfaceAnnotated<EthernetFrame>` to `EthernetFrame`.

As referenced above, this commit also introduces the
InterfaceAnnotated<Packet> structure as a means to demarkate which
interface a packet originated from, and where a packet should be sent
to. This information is needed at many points in the router to determine
the correct course of action for each packet.

Implementation Note: These links could have been implemented in a
generic fashion to faciliate more than the current three
interfaces (Host, LAN, WAN), but that would increase the complexity of
the link. It is not expected that this example will require more
interfaces, but if more are needed, supporting more interfaces shouldn't
be a onerous task.