PHR: Rework Intake and Exhaust Links
At our last meeting, other members brought up that we should not code
the IO layers into the business logic of our core router, particularly
since we want to be able to test the router using PCAPs, and swapping
real OS interfaces for PCAP streams shouldn't be a pain. So I am
reworking the intake and exhaust to not do this, and to handle
converting the raw Vec<u8> packet streams to route-rs Packets.