PHR: Classify packets based on inbound interface
This will allow us to make routing decisions based on whether packets
are coming from the lan or wan. Since the types allow us to receive
packets with an unmarked inbound interface, we drop them because we
don't know what to do. Include a test to document that this is intended.