PHR IPV4: Drop spurious packets from the LAN
Packets from the LAN should be destined for the Host (or broadcast) or
the WAN. Any packets that we receive destined for other LAN addresses
are an error, and we should drop them.

In order to implement this, we now have to know what our LAN IP and
subnet are, so we add fields and builders for those.

Note that we have to handle both the global broadcast 255.255.255.255
and the local broadcast which is (network | (broadcast & subnet mask)).
Fortunately, cidr provides .last_address() which does this for us.