Draft: AutoEgressor, auto unpack egressor vector
This is a super trivial commit using a macro to auto unpack a the
egressor vector into a generic length set of variables. I bet with some
more clever thinking we could make something even more ergonomic, but I
wanted to at least demonstrate this is possible. The macro is quite
simple, checkout rust_by_example variadic functions for more.