Packets: Allow chaining setters
Adjust the packets library so that setter functions may be chained. This
gives a more ergonomic pseudo-builder style, but doesn't require it,
since references that are uncollected are automatically dropped. I think
this does a good job providing a nice syntactic interface, without
forcing the user into a particular style. It also isn't a breaking
change to the interface.