Unpack!: A Macro to help you remove stuff from links
This is a preliminary cut, for a draft commit
It does two things at once:
For runnables, it takes the vector passed in, your `all_runnables` and
appends the runnables returned from the link to the provided vector. You
may also choose to provide a `_` instead, and the runnables are
discarded.

For egressors, if you provide a single variable, it is declared as your
would expect, and represents the vector of egressors that is returned
from the link. If you provide a one or more variables inside of `[]`
square brackets, it automatically destructures the vector into those the
provided variables.

Examples
unpack!( let (runnables, egressors) = myLink::new().build; );

unpack!( let (runnables, [a, b, c]) = myLink::new().build; );

unpack!( let (_, egressors) = myLink::new().build; );

unpack!( let (_, [a,b,c]) = myLink::new().build; );